feat(memory share): avoid updating PTs
Avoid updating PTs of the caller to retrieve request
and relinquish, when the caller is from the NWd.

In this case, simply update the share state structures.
S2 update for the endpoint should be managed by the
hypervisor.

Added a new entry in the `enum ffa_map_action` which is
used to represent the operations to be done to S2 PTs
during memory share/lend/donate, retrieve and relinquish.

If the retrieve request or the relinquish is from the NWd,
the functions `ffa_retrieve_check_update` and
`ffa_relinquish_check_update`, return MAP_ACTION_NONE to signal
the SPMC shall not operate the pages.

Change-Id: I78315e6b2684a19abe3749ddfedbc06579809693
Signed-off-by: J-Alves <joao.alves@arm.com>