feat(memory share): relinquish use `memcpy_trapped`
The handling of FFA_MEM_RELINQUISH copies the relinquish descriptor
from the TX buffer of the NWd for calls made from NWd VMs.
To safely do so, refactored the code copying the descriptor and
used `memcpy_trapped` to recover from a potential GPF. The copying
was split in two steps: first the relinquish descriptor and then
the rest of the message. The first step's intent is to safely
validate the content of the header before copying the whole descriptor.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Iad29ccfbb9df8e663a0f98aff41eea200a2a1a0d