fix: propagate enable_mte build flag to cflags
This patch makes sure only legal values, either 0 or 1, are specified
for enable_mte build flag from the project configuration file. The value
specified is propagated to cflags for compilation.

Further, march compiler flag is specified irrespective of whether the
enable_mte build flag is specified in the project configuration.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I2b09b6ac94b661533b669f6c98ffbb26deb49925