refactor: use `typeof` in `HFTEST_ASSERT_OP`
The `__typeof` keyword can be used instead of a union to declare
variables with unknown type. This simplifies the implementation of
`HFTEST_ASSERT`, and ensures that enums are not silently casted to ints.

The only drawback compared to the old method is that `typeof` cannot be
used on bitfields. But bitfields are rare, so we can fix all the uses by
loading the bitfield into a local variable first, then comparing it.

Change-Id: Ifbf4b70ccc261ef14f30bb5bb5a3c1c62b2a0cee
Signed-off-by: Karl Meakin <karl.meakin@arm.com>