fix(hftest): define stacks for all secondary cores
The hftest common libraries was reusing the same stack section
for all secondary cores.
This means some tests would fail unexpectidly.

Adding a function to common.c and service_common.c, which have
the same purpose in the target they are used in, to retrieve
the buffers for the secondary core stacks defined globally.

Signed-off-by: J-Alves <joao.alves@arm.com>
Signed-off-by: K-Meakin <karl.meakin@arm.com>
Change-Id: If93aed330a88c3b52e5771d09911296f2c6214ed