test: driver for generic memory mapped system timer
This patch introduces the framework for implementing a driver
for a generic memory mapped system timer periperal.

It will be used to implement the driver for AP REFCLK Generic timer
which will serve as a second source of secure interrupt.

Most of the code has been adopted from tf-a-tests, a sister project
in trustedfirmware.org

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: Id27fb5a0a249de02d567e00e6aac1d3b0e097489