test: leverage build define to identify an S-EL0 SP
Currently, any partition with single execution context is being
considered as an S-EL0 SP. This is not correct given the fact that
an S-EL1 SP can also be an UP SP.

In further patches, we introduce a test setup with S-EL1 UP SP as
service2. Hence, we leverage a build define to identify an S-EL0
partition.

Also, this patch relaxes the check for return status upon invocation
of FFA_SECONDARY_EP_REGISTER_64 ABI as it is expected to fail for an
S-EL1 UP SP.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I9be4d9bbca6a1e10ab40cb33f2743c130a342c01