test: interrupt to be pended before boot
SPs are booted according to boot-order field.

The first SP(SP1) to boot pends an interrupt in the eSPI
range, via a dedicated FVP platform-specific call to
pend eSPIs in EL3.

The 2nd SP(SP2) is assigned with a dummy peripheral,
only to assign it with the interrupt the 1st SP
will inject.

Test sequence:
- During runtime init, SP1 pends an eSPI for SP2.
- During runtime init, SP2 enables the eSPI, unmasks interrupts
through the PSTATE bits, and configures an interrupt handler.
- At this point, SP2 should handle the interrupt, setting
an internal flag for further validation in runtime.
- All SPs terminate booting according to boot-order, so execution
is given to the NWd.
- Test function executes SP2's service to check the eSPI
has been serviced in initialisation.

Test currently only applicable to MP S-EL1 partitions.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Icafcb5ee9cdab8a7246f4a2b16ac7c0e2e7ab291