feat(interrupts): physical interrupt enabled
SPs are assigned with the interrupts related to
peripherals in their specific FF-A manifest.

The SPMC was enabling the physical interrupt,
when loading the respective partition.

This is changed in this patch to only be done
later at the call "HF_INTERRUPT_ENABLE".
This signals the SPMC that the SP is ready
to handle the given interrupt.

This prevents an IRQ from being pending before the
SP was even able to prepare itself to handle it.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I8a1704b46e41464c24a191ef8b9d41b1183e1aff