fix: check `.h` files with `clang-tidy`
The `HeaderFilterRegex` regex in `.clang-tidy` was intended to ignore
headers under `third_party/`. But the regexes in `HeaderFilterRegex` do
not support negative lookaheads
(https://stackoverflow.com/questions/71797349/is-it-possible-to-ignore-a-header-with-clang-tidy),
so the result was that no headers matched the regex at all, and warnings
from any header files were suppressed!

When `clang-tidy-19` is released, we can use `ExcludeHeaderFilterRegex`.
Until then, list out explictly all the allowed paths.

Change-Id: Ifd182a0ca4558cae8dd1182933137d4a4caa3b6e
Signed-off-by: Karl Meakin <karl.meakin@arm.com>