fix(docs): point poetry readthedocs virtual env
RTD uses a mixture of poetry and pip to install packages in the build
environment. In the past it was recommended to disable poetry from
creating a fresh virtual environment. Instead, the expectation was that
poetry would be able to detect it's current virtual environment and
install the packages in the right place. This was recently updated to
allow poetry to better allow dependcy management by poetry [1]. Remove
this configuration and explicitly point Poetry to the virtual
environment.

[1] https://github.com/readthedocs/readthedocs.org/pull/11152

Signed-off-by: J-Alves <joao.alves@arm.com>
Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: I64e75410a0f14ff5edd91ea526d85ad108cb1a13