fix(docs): refactor poetry dependency group
Rename 'doc' group to 'docs' for consistency, this is to follow the
widely accepted convention of using plural nouns groups that contain
multiple items. This change signifies that the 'docs' group includes
a collection of documentation-related dependencies.

Also, ensure that the dependencies are actually conditionally
installed by setting the group as optional. This was missing in the
original change.

Change-Id: I33aa154d849a93bac78d698bbd2eef2f839e8985
Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>