fix: use correct load-address while adding offset
When a memory region is present in the SP manifest with
"load-address-relative-offset", load address is added to the offset to
get the actual address. The load address of a SP is mentioned in both
SPMC manifest and SP manifest. In case of a conflict between the two,
Hafnium will use the load address from SPMC manifest. Currently, this is
done after the memory regions mentioned in the SP manifest are parsed
using the load address mentioned in the SP manifest. This patch will
ensure that the actual address of the memory regions in the SP manifest
will be calculated using the load address from SPMC manifest and not
from SP manifest.

Made changes in the unit test as it is not feasible to use the load
address relative offset.

Change-Id: I864940445d6e5bd6a58978cf2c5b6a3f0c5f8cc3
Signed-off-by: Davidson K <davidson.kumaresan@arm.com>