refactor: reduce fields tracking interrupt handling for vcpus
With the changes made in this patch stack, there are alternatives
to detect the current virtual secure interrupt being handled by
a vcpu rather than tracking it explictly through a special field
in vcpu data structures. The first valid entry in the interrupt
queue provides this information.

With this in mind, we proceed with removing two of the fields related
to interrupt handling. In order to do so, we always add a virtual
secure interrupt to the queue irrespective of whether it is signaled
or not. This does not violate the spec and is simply an implementation
choice.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I7aef287d54493887d6eb9a72a71c40c29014494c