fix: remove memory region's device attribute
Memory and device region node may set attributes bit 4 to state a region
is a device region. This is bogus as despite not being mentioned in the
FF-A binding doc. [1], a memory region node may set device bit attribute
and result in defining a memory region as a device region.
Remove the device attribute from manifest parsing entirely as this comes
as a redundant information with respect to the way memory and device
region nodes are defined in the device tree format. This doesn't change
the parsing logic as memory and device regions are already segregated
into different data arrays.

[1] https://trustedfirmware-a.readthedocs.io/en/latest/components/ffa-manifest-binding.html#memory-regions

Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: If882a6b37bada1df33c2a06665071162ee138693