test(ipi): state machine to help testing IPI
Add a state machine to help test IPI support.

It tracks the transition through important phases,
from IPI enabled to the respective IRQ handler.

Refer to the doc comment in 'enum ipi_state' for more
details on each state.

The state tracking structure can be instantiated
in a shared buffer, such that the transition to sent
can be controlled externally to the target vCPU.
Which helps in testing that the IPI is sent, whilst
target vCPU is in the various states.

Signed-off-by: J-Alves <joao.alves@arm.com>
Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I0f506cd4c287a37f17c20ddaf76a16b4b013b7ba