test(ipi): handling SRI in the NWd
Introduce function to enable SRIs in the NWd as this will be
required for testing IPIs sent to a target vcpu where the
required SP is in the waiting state.

The changes to the build script mean that we need to move away from
using the HF_PRIMARY_VM_ID macro to the hf_vm_get_id() helper.

Signed-off-by: J-Alves <joao.alves@arm.com>
Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: If169ab680f1eb565e692c4bb52f6f968cbee9f8a