fix(hftest): clear NPI when polling for notifications
Before the service function is executed on SPs, the framework
gets the pending notifications from the framework and looks
for the RX full notification. The notification indicates
there is a new message in the mailbox.

When resuming the SP to handle the message, the SPMC pends
the NPI VI. As it can affect the running test, clear the NPI
as well when polling for notifications.

Signed-off-by: J-Alves <joao.alves@arm.com>
Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Change-Id: I4cb82daa6d10c960224a64994c6c4836f8567716