feat(arch timer): resume SP if deadline expires in NWd
Once a SP's vCPU is scheduled out and execution is returned to normal
world, it is likely that the timer deadline set by the vCPU could
expire in normal world.

It is the responsibility of SPMC to resume the SP's vCPU with expired
timer deadline and inject the timer virtual interrupt. It does so by
maintaining a list of all vcpus with pending timer and tracking the
nearest deadline with host timer whenever exiting to normal world.

When the SP's vCPU is eventually descheduled, SPMC will remove the vCPU
entry from the list if the corresponding arch timer is disabled.

Change-Id: I11d63f89cef320e41980eaf10188f8a72ed84cec
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>