feat(arch timer): handle spurious host timer interrupt
Whenever a spurious interrupt occurs due to S-EL2 host timer, simply
resume the preempted vCPU. One such scenario is described here:

SPMC on CPUx could be tracking a pending timer deadline of a SP's
vCPU. If NWd driver allocates CPU cycles to migrate this vCPU to
CPUy, the host timer on CPUx will still cause the interrupt to
trigger spuriously on CPUx. In this case, simply resume whichever
vCPU was preempted on CPUx. The S-EL2 host timer on that CPUy will
take care of signaling the timer virtual eventually.

Change-Id: I7118c4f0036d9eae95cc3512a54c98c4b6fd5959
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>