test: SP handles timer with short deadline
This patch adds a test that requests SP to start the timer with short
deadline and wait for few milliseconds for the deadline to. SPMC tracks
the timer deadline and signals the timer virtual interrupt which shall
be handled by the SP.

Further, the test queries the SP for the last serviced virtual interrupt
to ensure timer virtual interrupt was handled.

Change-Id: Iae5cbc8a44fc349a133a450da46ea3157e4124d5
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>