test: fork 'preempted_by_secure_interrupt'
Forked the test preempted_by_secure_interrupt test for both
the S-EL0 and S-EL1 partitions.

S-EL0 partitions need the hf_interrupt_deactivate call
after the FFA_MSG_WAIT call was preempted with an
eret FFA_INTERRUPT.

For S-EL1 partitions this was already done in the irq
handler.

To fork, added a new source file for the secure
interrupt-related services for S-EL0, which has the
same name as for S-EL1 services. This way, PVM can
still interact with both transparently.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Ic70c5b3cae49fc3071f84d2e17fd41d3eb3ca8ec