ci: drop the setup with the hafnium driver
The setup with the hafnium driver is making use of
legacy functionality which has been dropped from the
hafnium common code implementation.

Dropping the setup from our test script as it is
no longer meaningful, in the context of the SPMC,
current hafnium's main target.

As well, dropping the Linux submodule as it is presently
not needed in the CI, and all the dependencies to
linux build.

BREAKING: build at the current commit might be broken.
Intermidiate step to deprecating submodules
third_party/linux and the driver/linux.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Iffc9fe40c52e2554df44b77762d0cfc5acd2f918