chore: drop legacy timer support in hypervisor
In subsequent patches, we add support for arch timers in secure world.
Hafnium, as an SPMC, shall trap and emulate SP access to arch (EL1
physical) timer. This is very different from the legacy support in
Hafnium as a hypervisor.

Hence, to ensure the source code does not get too convoluted, the
legacy timer support in hypervisor is dropped.

BREAKING: Any tests that leverage EL1 physical timer as a source of
non-secure interrupt will fail. A new source of non-secure is
introduced in subsequent patches.

Change-Id: I5b61be53aeb710689464fe757c9cbf9a56c7b970
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>