test: introduce driver for sp805 peripheral
An SP805 peripheral can be instantiated to serve as a normal world
or secure watchdog timer source.

Hence, split the existing trusted watchdog driver to create a generic
sp805 driver. This will be helpful in creating a driver for normal
world watchdog timer which will act as a source of non-secure
interrupt.

Change-Id: I1b203fd5ec5d95ef720d2b7935a08803c8fd1481
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>