fix: simplify secure interrupt handling
Now that hafnium supports queueing of virtual interrupts for each
execution context of an SP, there is no need to disallow high priority
interrupts from preempting a vCPU while it is handling a virtual
interrupt.

Hence, remove the actions of masking and unmasking physical interrupts
before/after vCPU is run.

Arm recommends the priority for G1 non-secure interrupts to be in the
range 0x80 to 0xff when programming from secure world. Currently, the
SRI priority is incorrect. This patch also fixes the priority value.

Change-Id: Icdc36f3aa09fab83cb58214e14e5a9bd174bc28b
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>