refactor: always eret FFA_RUN to the caller
The FFA_RUN interface was not running an SP if it didn't
have any pending interrupts, notifications or messages.
This was inherited from the hypervisor implementation.

With the present change, if the vCPU is in VCPU_STATE_WAITING
always eret FFA_RUN to the target SP.

This behaviour aligns better with what is defined by the
runtime model for FFA_RUN interface in the FF-A specification
(FF-A v1.1 onwards).

BREAKING: this change the expected behaviour of the FFA_RUN
ABI. This breaks Hafnium tests, which are fixed in other
patches.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Ib37a430ef357720e7368fb4654b26f610986756e