ci: add script download checkpatch.pl
Hafnium's static code checks include those of linux
tree/scripts/checkpatch.pl.

Linux tree has been dropped from the submodules of
the hafnium project to trim the repo size and
the driver isn't that meaningful at this stage.

This patch adds a script to build/ to download the
checkpatch.pl from the linux v6.11, and set it up
in the out/scripts folder.

The Makefile is changed to:
- Find the script in the directory to where it is
downloaded, and added '--no-tree' to the arguments to
hint it shouldn't expect the linux tree.
- Extended the variable 'CHECKPATCH_IGNORE' which contains
a list of files to be excluded by the script. This is
because the checkpatch.pl complains about using __attribute__
which is allowed by clang, whilst the suggested alterantives
are not. Hence, the extension of the list is necessary.
- Invoke the "setup_checkpatch.sh" if "checkpatch.pl" is not
present.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: If2785b73c6d125d6b0c0e57ce8ed4b16dde9b10c