fix: only normal world VMs need FFA_RUN
In the SERVICE_SELECT macro, the services can either be SPs
or VMs.
SPs boot before the normal world. As such, the primary doesn't
need to invoke FFA_RUN for it to get to message loop.
Secondary VMs however, need to be bootstrapped by the Primary
VM.

This patch makes it such the FFA_RUN ABI is only used for
Secondary VMs.

Change-Id: I2346568fd453a876a0733359a6492311d051ed28
Signed-off-by: J-Alves <joao.alves@arm.com>