test: add helpers to share page for coordination btw endpoints
Various endpoints involved in certain test scenarios need to share
a page among themselves which acts as a buffer containing shared
variables.

The update made by one endpoint needs to be visible to other endpoints
involved in the test. The tests from IPI suite use this mechanism.

In subsequent patch, we leverage this mechanism to add a new test
scenario for secure interrupt handling.

Change-Id: I12deaf4f3b4eac717a8e288706a1d7044c0abecb
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>