test(ipi): test IPI to non primary CPU
Test that an IPI sent to a CPU other than CPU0 when the target
vCPU is in the waiting state will trigger and SRI to the normal
world and be correctly queued.

Add relinquish and reclaim step for the IPI state memory so
we can run the waiting test multiple times to test multiple cpus.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Iac81c375375904f4745aeb39702de7ea3d8175d3