refactor: maintain the boot order list with VMs
In the present implementation, the boot order is maintained through a
list of nodes contained by vCPU data structures.

Boot order protocol is a concept that naturally aligns with VMs. In
subsequent patches, we depend on the boot order for bootstrapping
execution contexts of MP endpoints on secondary CPUs. This motivates
us to maintain the boot order list with elements representing their
parent VM entries. This makes it simpler to obtain the vCPU to be
bootstrapped for initialization when a secondary CPU is turned on.

Change-Id: I333fa2b3188db99b2b52e7e43988a7a7b9fb7aef
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>