fix: fix static_checks.sh and `Makefile`
`static_checks.sh` interpreter directive was broken on MacOS, because
there is no file named `/bin/bash`. The portable way to specify bash as
a shell is `/usr/bin/env bash`.

`make license` was not running, because `make` checks if a file with the
same name as the target has changed, and if it is unchanged, doesn't run
the target.

`make tidy` was printing hundreds of useless "n warnings generated."
messages. Silence them.

Change-Id: I0d25cbed2dee61c6721bd2cca95fb981029ec48b
Signed-off-by: Karl Meakin <karl.meakin@arm.com>