fix: signal virtual interrupt to vCPU in waiting state
At the moment, SPMC queues the virtual interrupt targeting a vCPU that
has migrated to a different physical CPU and is waiting state.

This patch changes that behavior by letting the SPMC signal the virtual
interrupt to target vCPU and resuming on the current CPU where the
corresponding physical interrupt triggered.

Change-Id: I38f9e677d51ae293d02c4a355870c9aeded23e1c
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>