fix(notifications): counter pending notifications
The counter for the pending notifications in the system
was getting corrupted when setting that had been already
set.

Changes the unit test `vm_notifications_set_and_get` to
recreate state change.

Fixed the function `vm_notifications_state_set` to exclude
from the notifications bitmap, the bits which are already
set pending.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I8fa6e4e7190498fb4b389d68906d950a683b1136