test: send appropriate framework response function
For every function associated with VM availability framework messages,
there is an appropriate function response to acknowledge it. This
shall be sent back to the NWd as part of a framework direct response.

This patch also modifies the receiver SP's service function to not
expect the execution to be resumed after last valid response.

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I55372779fdbc306677fd31f8205444d8be372de9