refactor(interrupts): plat_ffa_find_target_vcpu
The function is used to obtain the target vCPU depending
on the interrupt.

Dealing with the special cases such as:
- IPI SGI.
- Architectural timer at S-EL2.
- Spurious interrupt ID.

Some of the above were dealt outside of the said function.
Which split the purpose of the function.

This way, the general interrupt handler is a easier
read as well.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I0e0731d353aef30c8327a726b3721c11bb53cd54