refactor(mm): move `stage1` into `mm_ptable`
Move the `stage1` flag from the `mm_flags` struct to the `mm_ptable`
struct. The flags struct describes options for each individual operation
on the page table, but the stage1/stage2 flag applies to the whole page
table (and doesn't change over the table's lifetime), so it should live
on the `mm_ptable` struct.

Change-Id: I98735b7267922b43bbc63a4f599ec410013d9220
Signed-off-by: Karl Meakin <karl.meakin@arm.com>