feat(interrupts): report CPU cycles to scheduler
Extend the FFA_NOTIFICATION_INFO_GET to include the
partition ID and vCPU IDs due to pending interrupt while
in the waiting state.

Drop the previous implementation of the same feature for
the IPI feature in Hafnium.
Partitions which need CPU due to IPI and interrupts due
to SRI policy configuration are reported in the same
function, and leveraging the same state structure,
to track whether the information has been retrieved via
FFA_NOTIFICATION_INFO_GET.

Change-Id: I39c1709dd7b63bf01c474cf4e7683f45e245908e
Signed-off-by: J-Alves <joao.alves@arm.com>