test(interupts): SRI triggered due to interrupt
Configured an SP with `sri-policy` to trigger SRI
when the SP is in waiting state, if it needs to be
resumed by the scheduler of the system due to
handle the interrupt.

In runtime, SP configures the trusted Watchdog
and sets itself in the Waiting state.

The PVM will look for the SRI, call FFA_NOTIFICATION_INFO_GET
and resume the SP.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Ia2f82359022f7985c5d293aebbc9753114ae28e2