test(interrupts): fix failing tests from interrupts change
The change to unify tracking of secure virtual interrupts means that
more NPIs are recieved by the SPs from indirect messaging used in test
coorination, these need to be correctly handled ensure the tests pass.

Also in the case of S-EL0 partitions since they do not receive the
interrupt through the vIRQ signal we treat the interrupt as no longer
pending after the FFA_INTERRUPT ERET so we do not need to call
hf_interrupt_get.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I5dcbe8feac1811ba3a723d74a34629c2c693667f