test(interrupts): SRI triggered into waiting state
Service3 is configured such that:
a) it is configured with sri-interrupts-policy to 2,
which means when it enters the waiting state with interrupts
pending, the SPMC should send an SRI to the NWd.
b) it is assigned with a dummy device whose interrupt
ID is on the ESPI range.

The ESPI assigned is used to test configuration a).

Service1 is configured with the ability to call platform
specific hooks to trigger an ESPI.

In this test, service3 executes with its interrupts disabled
and invokes FFA_MSG_WAIT.
From this point:
- SPMC sends SRI to the NWD.
- NWd handles SRI calling FFA_NOTIFICATION_INFO_GET.
- NWd calls FFA_RUN(service3, 0), so it can handle
ESPI.
- Service3 resumes, unmaskig interrupts. It is preempted
and the ESPI is handled.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I98e35f1a78fd3dc8a0ecdcdd02ca358cede5df13
Signed-off-by: J-Alves <joao.alves@arm.com>