feat(mm): report address of first mismatch
Rewrite `mm_get_attrs` to report the address of the first page that was
not mapped with the same attributes as the rest of the region.

This requires tracking a 2nd flag (whether a mismatch was found) as well
as a second return value (the address of the first mismatch), so replace
the parameters and return value with a `struct` encapsulating all the
state.

Change-Id: If397cd8a9cfa81bf0e600c60059ccea8540b9224
Signed-off-by: Karl Meakin <karl.meakin@arm.com>