refactor(interrupts): unify secure virtual interrupt tracking
Create an interface for setting and getting secure interrupts to
ensure that the bitmap and queue for tracking secure virtual
interrupts are kept consistent. This makes tracking secure interrupts
easier. As part of this make the queue manipulation functions private.

This requires use to clear the bitmap and pop from the queue
as part of HF_INTERRUPT_GET so remove popping from the queue
in HF_INTERRUPT_DEACTIVATE.

BREAKING: This change means that since all interrupts are now queued
some tests may break if they do not currently handle or disable any
interrupts (particularly NPI which is set during indirect messaging
used in test coordination).

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I5a8872fefcd0e79c6d9720033679f4505bc8c2eb