feat(interrupts): pop interrupt on FFA_INTERRUPT to S-EL0
S-EL0 partitions cannot receive interrupts through a vIRQ signal in
addition to the FFA_INTERRUPT eret and therefore will not clear the
interrupt in an interrupt handler. Because of this we mark the interrupt
as no longer pending at the point of the FFA_INTERRRUPT for S-EL0
partitions.

BREAKING: Any S-EL0 partitions that call hf_interrupt_get will fail
until fixed.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I9f51499c0fa1f9ee9bb3f905401784114e98e2c9