fix(gicv3): enable G1S interrupts if not enabled by TF-A
The GIC driver in TF-A programs the GICD settings for enabling G0
interrupts. But, the interrupts enabled in the SP manifests are
enabled by the SPMC (Hafnium). So, it makes sense to at least
enable the G1S bit in the GICD_CTLR register from Hafnium, if
it is not already enabled by TF-A.

This patch updates the gicv3_distif_init() routine to program this
bit.

Signed-off-by: Sungbae Yoo <sungbaey@nvidia.com>
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Change-Id: I0c256a2edb1392d6134f7320a723bf73ce3e2212