feat(interrupt): clear pending ME interrupts during msg_wait
Managed Exit is no longer required if the SP has returned to the
waiting state. So if there is a managed exit interrupt pending when
we are servicing a message wait or direct response request clear it.
vcpu_increment_clear_decrement is made static to ensure the bitmap,
queue and count are kept in sync.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Ibd1d41360c09054f8634a203ab09c09ea407415a