refactor(interrupts): make enable bitmap funcs private
To ensure that the interrupt count is kept in sync when enabling and
disabling secure interrupts make the functions that manipulate the
interrupt_enabled bitmap private.

This requires modifying the unit tests to use the
vcpu_virt_interrupt_enable function so use this opportunity to add a
TearDown function for the vCPU unit tests to make things cleaner.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: Ie707547677199837df023e964c751ffe93fdac9b