refactor(interrupts): use interrupt_resume_waiting
The function interrupt_resume_waiting is used to handle
secure interrupts when the target vCPU is in the waiting
state.

The operations to inject the interrupt when intercepting
the FFA_MSG_WAIT and FFA_MSG_SEND_DIRECT_RESP calls should
be similar. Given, the `ffa_interrupts_intercept_call`
assumes the handling of the operations has been taken care of
as if the caller vCPU didn't interrupts pending.

Then, when checking for interrupts pending and intercepting
the call, it can set the next vCPU to execute in the preempted
state, and resume calling vCPU in SPMC scheduled mode.

As such, slightly refactor the function
`interrupt_resume_waiting` to better cater for both
instances.

Change-Id: I0ffbb32c5b4325c76f2001a12febb92070cf1546
Signed-off-by: J-Alves <joao.alves@arm.com>