feat: abort vCPU upon encountering fatal exceptions
If an execution context of a S-EL0 partition encounters a synchronous or
asynchronous exception, it is considered fatal error and SPMC follows
the abort handling procedure outlined in the FF-A spec.

Note that S-EL1 partitions are still given an opportunity to handle the
exception by itself. If found to be fatal, the S-EL1 SP is expected to
invoke FFA_ABORT ABI explicitly to terminate execution.

Change-Id: I844a07961d6dbda1025e13c16f071b0accb4a690
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>