test: power down secondary CPUs while handling arch timer
This patch configures the architectural timers of the Service SPs to
periodically trigger at a higher frequency (i.e., shorter intervals)
to stress-test the handling of the CPU_OFF power management message by
the SPMC and/or SPs. The goal is to ensure that any interrupts,
generated while SPMC and/or SP is handling a power management operation,
do not interfere with such operations, as the SPMC is expected to mask
all interrupts during such procedures.

The verbosity of Service SP arch timer handler is changed to reduce
possibility of test timeouts due to excessive console output messages.

Change-Id: Id9be3ecbb0b5ce3ea7673f4b5d12c5be62cbeef4
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>