fix(ff-a): use cpu buffers in partition discovery
The handling of FFA_PARTITION_INFO_GET(_REGS) was using
a buffer for partition information.
The buffer was meant to hold the partition information
for all SPs - including logical SPs.

The buffer didn't cater for the support of multiple UUIDs.

Currently, there is the assumption that all information
should fit into a single RX buffer. CPU buffer is of the same
size as the mailbox.

Instead of statically defining a small buffer in the stack,
reuse the local buffer of the CPU handling the call.
The CPUs buffer are of the same size as the RX buffer.

Applied equivalent changes to both FFA_PARTITION_INFO_GET
and FFA_PARTITION_INFO_GET_REGS.
For FFA_PARTITION_INFO_GET_REGS, the return of information
is done via registers.
However, it still inherited the restrictions to
FFA_PARTITION_INFO_GET.

With the present approach, there is still the chance for
the limit of the buffer to be hit. However, current
buffer size is expect to suffice for known integrations.
In such case, the interfaces will simply return with an error.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I266262b5f3964693e983efb45603013e568e84b4