test: allow partition to signal initialization fail using FFA_ABORT
A Secure Partition can invoke FFA_ABORT to signal to SPMC that it
failed to initialize its execution context. SPMC is expected to
abort the partition but continue bootstrapping other partitions
based on the VM boot order.

In order to test above scenario, this patch refactors existing test
setup and makes the SP1 to make use of FFA_ABORT interface, instead of
FFA_ERROR, to signal initialization failure.

Change-Id: I07b8a761353eaba352989b37f31d4d833b273327
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>