fix(docker): env var HAFNIUM_FVP mounted
The HAFNIUM_FVP can be used to specify a custom location
of the FVP model, instead of the default one at
<Hafnium root>/../fvp.

Even though the hftest.py script was using it, and it was
passed to the development Docker image, the respective
directory was not mounted to the docker image.

This addition to the script mounts the location of
HAFNIUM_FVP binary if it is defined, otherwise
mounts the mentioned default.

Change-Id: I4e0fd8aef897d669b8009dc9d25349dd0ccd99d0
Signed-off-by: J-Alves <joao.alves@arm.com>