refactor(mm): drop ppool argument from init
Drop the ppool argument from the mm_init only.
The mm module is the most dependent on the ppool argument.
This defines a static 'ppool' inside the mm module.
Following patches will gradually drop the argument
to the functions which interact with external modules.

BREAKING: Build fails, to be reestablished at a later patch.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I1d19ea6c3614e3f572c860e5102ddd73410fa0d9