feat: reclaim memory regions of aborting VM
As part of abort handling, SPMC performs clean up action of reclaiming
memory regions on behalf of the aborting partition.

In order to do so, this patch refactors existing code to create helpers
that will be leveraged in the clean up actions.

This patch introduces a new deferred reclain flag that gets set when
a partition aborts but SPMC cannot immediatly reclaim the shared memory
region owned by the aborting partition because other borrowers have not
yet relinquished their access to it. This flag also ensures no partition
can ever succeed in retrieving the memory region originally owned by the
aborting partition.

Change-Id: I3a398bf9e64a60919ee3d4114504b956f1e0495d
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>