feat(memory_alloc): per CPU rollback memory allocation
For the memory management operations, the pages which are freed
during page tables update need to be preserved until the operation
is committed.
This is to allow the rollback of the operation in case it is needed.
So the allocator must ensure that memory freed during a rollback section
is held privately and not consumed by another allocator until the
operation completes.

As such, adding a rollback memory pool to the CPU structure. The memory pool
is meant to be initialized from the memory allocator API with the global
memory pool as the fallback to the local memory pool.
Then, when all memory can safely be relinquished, the cpu's local memory
pool can be destroyed and any memory tracked can be given to the global
memory pool.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I1c5340ae1d139be12c2dac0e69d233017adfd510